+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ***Added:***
 
 - Persist cache directories for the `linux-container` developer environment type
+- Add `clean_cache` and `cache_size` methods to the `linux-container` developer environment type
+- Add `env dev cache` command group
 
 ## 0.15.1 - 2025-06-05